Reorganized structure for separation of concerns and to simplify login toggle
