+++ MobileUI/contexts/LoginState.js
@@ -1,0 +1,4 @@
+import {createContext} from 'react';
+
+const LogStateContext = createContext(false);
+export default LogStateContext;
