+++ MobileUI/contexts/User.js
@@ -1,0 +1,4 @@
+import {createContext} from 'react';
+
+const UserContext = createContext();
+export default UserContext;
